requirements.txt

faiss-cpu>=1.10.0
fastapi>=0.115.12
httpx>=0.28.1
openai>=1.70.0
python-dotenv>=1.1.0
pandas>=2.2.0
uvicorn>=0.34.0
pydantic>=2.6.1
requests>=2.32.3
numpy>=1.26.0
tiktoken
rapidfuzz
ussd-session
asyncpg>=0.29.0
